+++ src/main/scala/info/gratour/jt808common/spi/model/TermCmdCompleted.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- *  Copyright (c) 2019, 2020 lucendar.com.
+ *  Copyright (c) 2019, 2022 lucendar.com.
  *  All rights reserved.
  *
  *  Contributors:
@@ -13,23 +13,26 @@ import info.gratour.jt808common.protocol.msg.types.ackparams.JT808AckParams;
 import java.lang.reflect.Type;
 import java.util.StringJoiner;
 
-public class CmdAck {
+public class TermCmdCompleted {
 
-    public static final Type TYPE = new TypeToken<CmdAck>(){}.getType();
+    public static final Type TYPE = new TypeToken<TermCmdCompleted>(){}.getType();
 
     private String id;
-    private String initiateMsgId;
-    private String subCmdTyp;
-    private String simNo;
     private long reqTm;
+
+    private int status;
+    private long tm;
+
+    private String initMsgId;
+    private String subCmdTyp;
+
+    private String simNo;
     private String plateNo;
-    private int plateColor;
+    private Integer plateColor;
+
     private String ackMsgId;
     private Integer ackSeqNo;
-    private long ackTm;
-    private boolean ok;
     private Integer ackCode;
-    private int cmdSrc;
     private JT808AckParams ackParams;
 
     public String getId() {
@@ -40,19 +43,36 @@ public class CmdAck {
         this.id = id;
     }
 
-    public String getInitiateMsgId() {
-        return initiateMsgId;
+    public long getReqTm() {
+        return reqTm;
     }
 
-    public void setInitiateMsgId(String initiateMsgId) {
-        this.initiateMsgId = initiateMsgId;
+    public void setReqTm(long reqTm) {
+        this.reqTm = reqTm;
     }
 
-    public int initiateMsgIdToInt() {
-        if (initiateMsgId != null)
-            return Integer.parseInt(initiateMsgId, 16);
-        else
-            return 0;
+    public int getStatus() {
+        return status;
+    }
+
+    public void setStatus(int status) {
+        this.status = status;
+    }
+
+    public long getTm() {
+        return tm;
+    }
+
+    public void setTm(long tm) {
+        this.tm = tm;
+    }
+
+    public String getInitMsgId() {
+        return initMsgId;
+    }
+
+    public void setInitMsgId(String initMsgId) {
+        this.initMsgId = initMsgId;
     }
 
     public String getSubCmdTyp() {
@@ -71,14 +91,6 @@ public class CmdAck {
         this.simNo = simNo;
     }
 
-    public long getReqTm() {
-        return reqTm;
-    }
-
-    public void setReqTm(long reqTm) {
-        this.reqTm = reqTm;
-    }
-
     public String getPlateNo() {
         return plateNo;
     }
@@ -87,18 +99,14 @@ public class CmdAck {
         this.plateNo = plateNo;
     }
 
-    public int getPlateColor() {
+    public Integer getPlateColor() {
         return plateColor;
     }
 
-    public void setPlateColor(int plateColor) {
+    public void setPlateColor(Integer plateColor) {
         this.plateColor = plateColor;
     }
 
-    /**
-     * Get message id of the ack(not the initiate message id).
-     * @return message id of the ack
-     */
     public String getAckMsgId() {
         return ackMsgId;
     }
@@ -106,14 +114,6 @@ public class CmdAck {
     public void setAckMsgId(String ackMsgId) {
         this.ackMsgId = ackMsgId;
     }
-
-    public int ackMsgIdToInt() {
-        if (ackMsgId != null)
-            return Integer.parseInt(ackMsgId, 16);
-        else
-            return 0;
-    }
-
 
     public Integer getAckSeqNo() {
         return ackSeqNo;
@@ -123,41 +123,12 @@ public class CmdAck {
         this.ackSeqNo = ackSeqNo;
     }
 
-    /**
-     * Ack time in epoch millis.
-     *
-     * @return ack time in epoch millis.
-     */
-    public long getAckTm() {
-        return ackTm;
-    }
-
-    public void setAckTm(long ackTm) {
-        this.ackTm = ackTm;
-    }
-
-    public boolean isOk() {
-        return ok;
-    }
-
-    public void setOk(boolean ok) {
-        this.ok = ok;
-    }
-
     public Integer getAckCode() {
         return ackCode;
     }
 
     public void setAckCode(Integer ackCode) {
         this.ackCode = ackCode;
-    }
-
-    public int getCmdSrc() {
-        return cmdSrc;
-    }
-
-    public void setCmdSrc(int cmdSrc) {
-        this.cmdSrc = cmdSrc;
     }
 
     public JT808AckParams getAckParams() {
@@ -170,20 +141,19 @@ public class CmdAck {
 
     @Override
     public String toString() {
-        return new StringJoiner(", ", CmdAck.class.getSimpleName() + "[", "]")
+        return new StringJoiner(", ", TermCmdCompleted.class.getSimpleName() + "[", "]")
                 .add("id='" + id + "'")
-                .add("initiateMsgId='" + initiateMsgId + "'")
+                .add("reqTm=" + reqTm)
+                .add("status=" + status)
+                .add("tm=" + tm)
+                .add("initMsgId='" + initMsgId + "'")
                 .add("subCmdTyp='" + subCmdTyp + "'")
                 .add("simNo='" + simNo + "'")
-                .add("reqTm=" + reqTm)
                 .add("plateNo='" + plateNo + "'")
                 .add("plateColor=" + plateColor)
                 .add("ackMsgId='" + ackMsgId + "'")
                 .add("ackSeqNo=" + ackSeqNo)
-                .add("ackTm=" + ackTm)
-                .add("ok=" + ok)
                 .add("ackCode=" + ackCode)
-                .add("cmdSrc=" + cmdSrc)
                 .add("ackParams=" + ackParams)
                 .toString();
     }